loaddata fixed
For an unknown reason, loaddata started producing an error. Error seems to be derived from a stock not having the full amount of time steps as the rest of the stocks. this made it so when reinitializing the array to hold data for a given stock, it was too small. This has been fixed with variable timesteps.